requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
imbalanced-learn==0.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
numpy==1.19.1
pandas==1.0.5
protobuf==3.13.0
pycaret==2.0.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wincertstore==0.2
xgboost==0.90
